requirements.txt
Flask>=2.3.0
Flask-Session>=0.5.0
python-dotenv>=1.0.1
PyMuPDF>=1.23.21
unidecode>=1.3.6
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.2
langchain>=0.1.14
langchain-core>=0.1.39
langchain-groq>=0.0.3